requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
board==1.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.7
cssselect2==0.7.0
defusedxml==0.7.1
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-htmx==1.21.0
djangorestframework==3.15.2
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
kaleido==0.2.1
loguru==0.7.2
multidict==6.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pluggy==1.5.0
propcache==0.2.0
psutil==6.1.0
psycopg2==2.9.10
pycparser==2.22
pyperclip==1.9.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
tenacity==9.0.0
tinycss2==1.4.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
webencodings==0.5.1
win32-setctime==1.1.0
yarl==1.18.0
